added orders controller

+++ src/main/java/spring/mvc/dao/OrderDao.java
@@ -17,7 +17,7 @@ public interface OrderDao {
 
     List findAllClosed();
 
-    Orders findById(Integer id);
+    Orders findById(Long id);
 
     void removeAll();
 }
